+++ R/np.semihat.R
@@ -821,6 +821,7 @@
     coef = coef,
     theta = theta,
     ridge = solve.out$ridge,
+    invalid = solve.out[["invalid"]],
     tww = tww
   )